addNewCard fix type of id

+++ src/main/java/cpe/atelier2/repository/card/CardJpaRepository.java
@@ -2,6 +2,6 @@ package cpe.atelier2.repository.card;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardJpaRepository extends JpaRepository<CardEntity, Integer> {
+public interface CardJpaRepository extends JpaRepository<CardEntity, Long> {
     CardEntity findByName(String name);
 }
